chore(suite-native): rename view only dir
