+++ suite-native/module-settings/src/components/EjectWallets/AutoEjectSwitch.tsx
