+++ suite-native/module-settings/src/components/EjectWallets/DevicesManagement.tsx
