+++ suite-native/module-settings/src/components/EjectWallets/WalletRememberModeIconButton.tsx
