+++ suite-native/module-settings/src/components/EjectWallets/WalletRow.tsx
